move programming exercise plagariasm check to dedicated service
